requirements.txt
anyio==4.0.0
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
Click==7.0
cryptography==41.0.0
exceptiongroup==1.1.3
future==0.18.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==2.8
pathspec==0.6.0
pycparser==2.19
python-telegram-bot==20.6
regex==2019.12.20
requests==2.31.0
six==1.13.0
sniffio==1.3.0
toml==0.10.0
tornado==6.3.3
typed-ast==1.5.5
urllib3==1.26.17
